feat: 피드백 반영하여 수정
