+++ src/main/java/ladder/domain/GameResult.java
@@ -11,6 +11,6 @@ public class GameResult {
 
     public String getResultName(int playerIdx) {
         int resultIdx = ladder.findResultIdx(playerIdx);
-        return results.getName(resultIdx);
+        return results.getNameAt(resultIdx);
     }
 }
