+++ src/main/java/ladder/domain/Player.java
@@ -21,6 +21,10 @@ public class Player {
         return name;
     }
 
+    public boolean hasName(String name) {
+        return this.name.equals(name);
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
